refactor: Repository에 있는 코드 Cacher로 옮기기
